+++ tests/TypeInspectors/IsListTest.php
@@ -47,6 +47,8 @@ use ArrayObject;
 use Closure;
 use stdClass;
 use Traversable;
+use GanbaroDigital\Defensive\V1\Interfaces\Check;
+use GanbaroDigital\Defensive\V1\Interfaces\ListCheck;
 use GanbaroDigital\MissingBits\TypeInspectors\IsList;
 use PHPUnit_Framework_TestCase;
 
@@ -201,6 +203,95 @@ class IsListTest extends PHPUnit_Framework_TestCase
 
         // ----------------------------------------------------------------
         // test the results
+    }
+
+    /**
+     * @coversNothing
+     */
+    public function test_is_Check()
+    {
+        // ----------------------------------------------------------------
+        // setup your test
+
+        // ----------------------------------------------------------------
+        // perform the change
+
+        $unit = new IsList;
+
+        // ----------------------------------------------------------------
+        // test the results
+
+        $this->assertInstanceOf(Check::class, $unit);
+    }
+
+    /**
+     * @covers ::inspect
+     */
+    public function test_can_be_used_as_Check()
+    {
+        // ----------------------------------------------------------------
+        // setup your test
+
+        $unit = new IsList;
+
+        // ----------------------------------------------------------------
+        // perform the change
+
+        $actualResult1 = $unit->inspect([]);
+        $actualResult2 = $unit->inspect(STDIN);
+
+        // ----------------------------------------------------------------
+        // test the results
+
+        $this->assertTrue($actualResult1);
+        $this->assertFalse($actualResult2);
+    }
+
+    /**
+     * @coversNothing
+     */
+    public function test_is_ListCheck()
+    {
+        // ----------------------------------------------------------------
+        // setup your test
+
+        // ----------------------------------------------------------------
+        // perform the change
+
+        $unit = new IsList;
+
+        // ----------------------------------------------------------------
+        // test the results
+
+        $this->assertInstanceOf(ListCheck::class, $unit);
+    }
+
+    /**
+     * @covers ::checkList
+     * @covers ::inspectList
+     */
+    public function test_can_be_used_as_ListCheck()
+    {
+        // ----------------------------------------------------------------
+        // setup your test
+
+        // ----------------------------------------------------------------
+        // perform the change
+
+        $actualResult1 = IsList::checkList([
+            [ 1 ],
+            [ 2 ]
+        ]);
+        $actualResult2 = IsList::checkList([
+            1,
+            2
+        ]);
+
+        // ----------------------------------------------------------------
+        // test the results
+
+        $this->assertTrue($actualResult1);
+        $this->assertFalse($actualResult2);
     }
 
     /**