+++ docs/changelog.md
@@ -28,6 +28,7 @@ pageflow_next_text: Contributing
 * Added `Check` and `ListCheck` support to existing classes
   - added `IsStringy::check`
   - `is_stringy()` now uses `IsStringy::check()`
+  - `IsList` now implements `Check` and `ListCheck`
 
 ## v1.9.0